requirements.txt
bert-extractive-summarizer

spacy==2.1.3
transformers==2.2.2
neuralcoref
